requirements.txt
Flask==2.2.2
PyJWT==2.7.0
pymongo==4.3.3